clean and add datasets
clean and add datasets

add engineering data

remove some raw data

add tests

rename demos to simulations
