+++ autoemulate/datasets.py
@@ -7,90 +7,40 @@ from sklearn.model_selection import train_test_split
 data_dir = Path(__file__).parent.parent / "data"
 
 
-def fetch_cardiac_data(dataset_name, train_test=False, test_size=0.2, random_state=42):
+def fetch_data(dataset, split=False, test_size=0.2, random_state=42):
     """
-    Fetches a dataset by name.
+    Fetch a dataset by name.
 
     Parameters
     ----------
-    dataset_name : str
-        The name of the dataset to fetch. Must be one of "atrial_cell", "four_chamber", or "circ_adapt".
-        All data from here: https://zenodo.org/records/7405335
-        "atrial_cell" is the ionic atrial cell model data from LH sampling.
-        "four_chamber" is the 4-chamber model without LH sampling.
-        "circ_adapt" is the circulatory adaptation model also from LH sampling.
-    train_test : bool, optional
-        If True, returns the dataset split into training and testing sets,
-        X_train, X_test, y_train, y_test.
-        If False, returns the entire dataset. Default is False.
+    dataset : str
+        cardiac1: ionic atrial cell model data from LH sampling.
+        cardiac2: isotonic contraction ventricular cell model, no LH sampling.
+        cardiac3: CircAdapt: four-chamber pressure and volume CircAdapt ODE model from LH sampling.
+        cardiac4: four chamber: 3D-0D four-chamber electromechanics model to predict pressure and volume biomarkers for cardiac function.
+        cardiac5: passive mechanics: inflated volumes and mean atrial and ventricular fiber strains for a passive inflation.
+        cardiac6: tissue electrophysiology: predict total atrial and ventricular activation times with an Eikonal model.
+        climate1: GENIE model: predict climate variables SAT, ACC, VEGC, SOILC, MAXPMOC, OCN_O2, fCaCO3, SIAREA_S.
+        engineering1: Cantilever truss simulation.
+    split : bool, optional
+        Whether to split the data into training and testing sets. Default is False.
     test_size : float, optional
-        The proportion of the dataset to include in the test split.
-        Only used if train_test is True. Default is 0.2.
+        The proportion of the dataset to include in the test split. Default is 0.2.
     random_state : int, optional
-        Controls the randomness of the dataset split.
-        Only used if train_test is True. Default is 42.
+        Controls the shuffling applied to the data before applying the split. Default is 42.
 
     Returns
     -------
-    X : ndarray, shape (n_samples, n_features)
-        Input data.
-    y : ndarray, shape (n_samples, n_outputs)
-        Output data.
+    X : array-like
+        Simulation parameters / inputs.
+    y : array-like
+        Simulation outputs.
     """
-
-    data_dir_dataset = data_dir / dataset_name
+    data_dir_dataset = data_dir / dataset / "processed"
     X = pd.read_csv(data_dir_dataset / "parameters.csv").to_numpy()
     y = pd.read_csv(data_dir_dataset / "outputs.csv").to_numpy()
 
-    if train_test:
-        return train_test_split(X, y, test_size=test_size, random_state=random_state)
-    else:
-        return X, y
-
-
-def fetch_cantilever_data(
-    dataset_name, train_test=False, test_size=0.2, random_state=42
-):
-    """
-    Fetches a dataset by name.
-
-    Parameters
-    ----------
-    dataset_name : str
-        The name of the dataset to fetch. Must be one of "cantilever" or "cantilever_truss".
-        All data from Zack Conti (Turing).
-        "cantilever": a dataset (996 dat points, input space is pseudo-randomly sampled) from
-        the finite element analysis (frame) of a very simple cantilever
-        beam where span of the beam, and it's cross-sectional depth are the inputs while
-        maximum deflection and weight are the targets/outputs.
-        "cantilever_truss": a more intricate cantilever truss design example whose
-        geometry is parameterised by a set of geometric variables. In this dataset,
-        deflection and weight are the targets/outputs while the remaining variables are
-        inputs, controlling the geometry. Dataset size: 1000.
-    train_test : bool, optional
-        If True, returns the dataset split into training and testing sets,
-        X_train, X_test, y_train, y_test.
-        If False, returns the entire dataset. Default is False.
-    test_size : float, optional
-        The proportion of the dataset to include in the test split.
-        Only used if train_test is True. Default is 0.2.
-    random_state : int, optional
-        Controls the randomness of the dataset split.
-        Only used if train_test is True. Default is 42.
-
-    Returns
-    -------
-    X : ndarray, shape (n_samples, n_features)
-        Input data.
-    y : ndarray, shape (n_samples, n_outputs)
-        Output data.
-    """
-
-    data_dir_dataset = data_dir / dataset_name
-    X = pd.read_csv(data_dir_dataset / "parameters.csv").to_numpy()
-    y = pd.read_csv(data_dir_dataset / "outputs.csv").to_numpy()
-
-    if train_test:
+    if split:
         return train_test_split(X, y, test_size=test_size, random_state=random_state)
     else:
         return X, y
