+++ data/cardiac2/description.md
@@ -1,0 +1,11 @@
+# cardiac2
+
+* data: https://zenodo.org/records/7405335  
+    * download `ventricular_cell_model`
+    * directory: `ventricular_cell_model/contraction_isotonic`
+* details: https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1011257
+* contact: Marina Strocchi / Steven Niederer
+
+### Description:
+* ToR-ORd model coupled with the Land contraction model samples used to train GPEs to predict the ventricular active tension transient features. The simulations were isotonic.
+* not sampled with LHS
