+++ data/cardiac3/description.md
@@ -1,0 +1,11 @@
+# cardiac3
+
+* data: https://zenodo.org/records/7405335  
+    * download `CircAdapt.tar.xz`
+    * directory: `CircAdapt`
+* details: https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1011257
+* contact: Marina Strocchi / Steven Niederer
+
+### Description:
+* Predicts four-chamber pressure and volume features with the CircAdapt ODE model.
+* sampled with LHS
