+++ data/cardiac4/description.md
@@ -1,0 +1,11 @@
+# cardiac4
+
+* data: https://zenodo.org/records/7405335  
+    * download `fourchamber.tar.xz`
+    * directory: `fourchamber`
+* details: https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1011257
+* contact: Marina Strocchi / Steven Niederer
+
+### Description:
+* samples generated with a 3D-0D four-chamber electromechanics model to predict pressure and volume biomarkers for cardiac function.
+* non-LHS sampled
