+++ data/cardiac6/description.md
@@ -1,0 +1,11 @@
+# cardiac6
+
+* data: https://zenodo.org/records/7405335  
+    * download `tissue_electrophysiology.tar.xz`
+    * directory: `tissue_electrophysiology`
+* details: https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1011257
+* contact: Marina Strocchi / Steven Niederer
+
+### Description:
+* predict total atrial and ventricular activation times with an Eikonal model.
+* LHS sampled
