+++ data/cardiac6/raw/outputs.csv
@@ -1,0 +1,181 @@
+"TAT,a","TAT,v"
+174.753,113.729
+221.656,152.993
+207.649,182.011
+345.776,261.984
+157.687,165.22
+206.646,276.821
+206.265,121.59
+155.32,320.299
+252.557,182.312
+248.854,112.889
+326.647,138.432
+139.227,187.659
+324.521,309.51
+422.187,125.347
+327.544,229.424
+176.073,161.981
+443.556,114.373
+131.133,232.445
+484.437,145.057
+148.267,341.046
+159.006,262.454
+261.141,120.768
+177.362,157.686
+138.32,171.693
+185.607,328.024
+534.349,233.616
+266.612,114.415
+213.366,137.589
+140.043,137.182
+375.615,191.115
+258.86,148.396
+274.868,99.6494
+185.291,128.59
+200.156,206.362
+139.187,271.416
+198.462,140.459
+568.908,175.391
+133.09,187.021
+366.595,140.004
+164.986,199.448
+148.141,231.079
+625.403,110.045
+384.168,113.022
+468.997,139.779
+292.425,183.592
+147.155,191.106
+500.647,273.192
+296.193,270.228
+354.703,208.831
+182.575,206.158
+200.341,210.523
+257.646,171.405
+186.185,134.276
+364.139,299.002
+224.923,260.966
+374.741,135.297
+211.572,245.25
+269.145,161.129
+212.413,139.299
+187.708,206.698
+238.294,161.218
+415.709,159.616
+187.673,126.138
+194.112,130.082
+284.09,105.513
+172.866,280.997
+176.076,227.213
+214.994,128.905
+291.165,227.536
+254.826,158.309
+231.81,194.034
+150.925,216.311
+164.6,153.105
+116.685,135.599
+182.161,175.963
+153.691,207.536
+320.541,230.203
+425.381,125.453
+196.892,106.857
+188.176,146.76
+138.053,195.382
+131.934,249.76
+207.359,226.679
+125.735,207.625
+430.236,267.208
+203.734,216.146
+274.813,249.82
+164.888,288.003
+171.154,238.336
+325.951,158.236
+247.634,148.699
+130.923,172.481
+187.501,175.475
+142.395,272.368
+572.117,175.663
+384.304,112.719
+208.93,144.866
+146.837,151.272
+192.788,166.597
+299.622,224.365
+341.139,225.765
+210.281,226.315
+396.032,164.377
+168.025,161.284
+500.85,251.896
+208.293,159.093
+150.343,206.175
+183.505,221.423
+129.479,237.315
+287.159,175.232
+225.034,135.882
+133.606,226.827
+381.713,251.518
+260.46,109.203
+190.034,123.969
+137.302,138.081
+217.896,232.916
+157.597,142.23
+145.479,151.521
+288.387,105.404
+169.642,120.529
+289.272,282.515
+312.324,227.712
+204.377,280.654
+140.249,347.809
+216.983,170.194
+225.936,117.77
+134.219,174.388
+156.613,199.1
+427.519,152.369
+239.524,202.095
+133.85,127.961
+289.897,166.671
+165.015,181.076
+286.226,147.044
+277.698,201.313
+188.15,219.763
+363.864,152.795
+187.159,153.809
+214.288,317.268
+323.73,111.475
+149.08,222.248
+113.612,147.317
+176.952,138.259
+220.405,138.313
+305.541,274.774
+255.459,150.704
+171.284,206.344
+154.915,231.485
+165.096,281.309
+177.25,269.665
+198.411,169.279
+203.128,241.012
+223.869,111.074
+144.059,386.992
+231.257,131.082
+273.529,263.862
+260.439,291.131
+284.539,190.894
+189.17,200.46
+188.202,182.373
+183.777,240.882
+321.667,176.829
+221.672,188.708
+183.111,106.165
+151.767,250.427
+447.711,259.715
+154.761,136.573
+143.816,266.495
+324.677,177.949
+383.527,182.54
+230.762,143.261
+230.86,144.162
+206.311,247.094
+194.636,122.871
+277.979,133.372
+458.691,194.972
+169.664,212.073
+312.749,172.112
+265.588,177.973
